As suggested by autoreconf, keep libool macros in m4 directory.

+++ bootstrap.sh
@@ -1,7 +1,13 @@
 #! /bin/sh
 
-rm -rf config
+if [[ -d config ]]; then
+	rm -rf config
+fi
 mkdir -p config
+if [[ -d m4 ]]; then
+	rm -rf m4
+fi
+mkdir -p m4
 
 #aclocal \
 #&& libtoolize --force --copy \
